Fix broken link on landing page
Link to the liveCOMS journal on the landing page was broken.

Included `https://` seems to be fix. Apparently without that,
the link is treated as a relative instead of an absolute link.

Was not able to test on my machine however, but that seems to be the
case for the other working links to LiveCOMS throughout the site.