Suggestion re out-of-date reviews
For the typical author, signing on for a perpetual review may seem like a long-term burden.  I wanted to make it clear that submitting to Livecoms and then not updating would still yield a publication in the same way a regular review would.  We don't want to dis-incentivize too much!