Updated enumerated lists:
- Changed upper-most enumerated levels to second-level titles.
- Shifted enumeration levels one level up (since upper-most levels are
  not enumerated lists anymore, but headers).
- Moved enumeration css-snippet into general page css.
- Some whitespace changes to unify indentation levels.

This has the down-side of not having automatic enumeration anymore
(need to change titles by hand if a new subtitle is inserted), but
has a number of advantages, namely that they look like headers
(there does not seem to be a way to do that in standard Markdown /
Jekyll otherwise), and that the enumeration scheme (1. / A. / i. / a.)
is now general enough to move it to a general css file.